fix(trir): key the free-variable cache by chunk identity, not by address
The per-chunk cache was keyed on `&TrChunk as usize`. A dropped
`CompiledFunction` — an `EVAL`, an on-the-fly compile — frees its chunk, and
the allocator may hand the same address to the next one, which would silently
serve one routine's cached free-variable bindings to another. `TrChunk` now
carries a monotonic id and the cache is keyed on that.

Refs #8895

+++ src/runtime/mod.rs
@@ -3767,7 +3767,10 @@ pub struct Interpreter {
     /// and validated against [`Interpreter::unit_lexical_gen`]; resolving
     /// them per call cost three string-keyed hash lookups plus two
     /// thread-local interner hits, which is more than `nom-ws`'s whole body.
-    pub(crate) trir_outer_cache: rustc_hash::FxHashMap<usize, (u64, Vec<Value>)>,
+    /// Keyed by [`crate::trir::TrChunk::id`] — a monotonic counter, not the
+    /// chunk's address, which the allocator may reuse after an `EVAL`'s
+    /// compiled routines are dropped.
+    pub(crate) trir_outer_cache: rustc_hash::FxHashMap<u64, (u64, Vec<Value>)>,
     /// Current frame's captured upvalue array, indexed by the running
     /// `CompiledCode::upvalue_syms` order. Read by `GetUpvalue(i)`. Set from
     /// `SubData::upvalues` on closure entry and saved/restored across call frames
